+++ Timer.DAL/Entities/Record.cs
@@ -38,7 +38,7 @@ namespace GtdTimerDAL.Entities
         /// <summary>
         /// Gets or sets how much time this record was runned
         /// </summary>
-        public int ElapsedTime { get; set; }
+        public double ElapsedTime { get; set; }
 
         /// <summary>
         /// Gets or sets watch type column
